arch-riscv: change fence's opType
Originally, the fence instruction will be dispatched to mem's dispatchQueue,
but its opType is No_OpClass, which will cause it to wait for the integer issue queue IQ2(IntMisc) to have free items before it can continue execution.

If the subsequent instructions of the fence instruction occupy the intIQ2, the fence cannot be executed and cpu stucks.

Therefore, change the opType of the fence instruction to MemReadOp to prevent this situation (in fact, the fence will not be dispatched to IQ)

Change-Id: Ie38a901e038db9906c43f78675e69391e847c88b